Move grpc client
