+++ orders/config.yaml
@@ -2,3 +2,5 @@ DB_URIS:
     "orders:Base": postgresql://${DB_USER:postgres}:${DB_PASSWORD:password}@${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:orders}
 
 AMQP_URI: amqp://${RABBIT_USER:guest}:${RABBIT_PASSWORD:guest}@${RABBIT_HOST:localhost}:${RABBIT_PORT:5672}/
+
+GRPC_BIND_PORT: ${GRPC_BIND_PORT:50052}
