+++ products/config.yaml
@@ -1,3 +1,5 @@
 AMQP_URI: amqp://${RABBIT_USER:guest}:${RABBIT_PASSWORD:guest}@${RABBIT_HOST:localhost}:${RABBIT_PORT:5672}/
 
 REDIS_URI: redis://user:${REDIS_PASSWORD:password}@${REDIS_HOST:localhost}:${REDIS_PORT:6379}/${REDIS_INDEX:11}
+
+GRPC_BIND_PORT: ${GRPC_BIND_PORT:50051}
